+++ libs/designer/src/lib/core/utils/openapi/schema.ts
@@ -1,4 +1,5 @@
 import { ExtensionProperties as SwaggerExtensionProperties } from '@microsoft/parsers-logic-apps';
+import type { OpenAPIV2 } from '@microsoft/utils-logic-apps';
 
 export const getTitleOrSummary = (schema: OpenAPIV2.SchemaObject): string | undefined => {
   if (!schema) {