+++ libs/designer/src/lib/core/utils/parameters/dynamicdata.ts
@@ -56,7 +56,7 @@ import {
   SchemaProcessor,
   WildIndexSegment,
 } from '@microsoft/parsers-logic-apps';
-import type { Connection, Connector, OperationInfo, OperationManifest } from '@microsoft/utils-logic-apps';
+import type { Connection, Connector, OpenAPIV2, OperationInfo, OperationManifest } from '@microsoft/utils-logic-apps';
 import {
   first,
   getObjectPropertyValue,
@@ -447,7 +447,6 @@ async function getManagedIdentityRequestProperties(
 
   let managedIdentityRequestProperties: ManagedIdentityRequestProperties | undefined;
 
-  // TODO - Update this when support for Managed identity is added.
   if (isManagedIdentityTypeConnection) {
     managedIdentityRequestProperties = {
       connection: { id: connection.id },